final_project_weather_api/requirements.txt
requests==2.31.0
pandas==2.0.3
psycopg2-binary==2.9.9